chore: avoid duplicate types causing _2 wip
